+++ next/cmd/update.go
@@ -1,15 +1,19 @@
 package cmd
 
 import (
+	"errors"
+
+	"github.com/go-git/go-git/v5"
 	"github.com/spf13/cobra"
 
 	"github.com/twpayne/chezmoi/next/internal/chezmoi"
 )
 
 type updateCmdConfig struct {
-	apply     bool
-	include   *chezmoi.IncludeSet
-	recursive bool
+	apply         bool
+	include       *chezmoi.IncludeSet
+	recursive     bool
+	useBuiltinGit bool
 }
 
 func (c *Config) newUpdateCmd() *cobra.Command {
@@ -31,13 +35,39 @@ func (c *Config) newUpdateCmd() *cobra.Command {
 	persistentFlags.BoolVarP(&c.update.apply, "apply", "a", c.update.apply, "apply after pulling")
 	persistentFlags.VarP(c.update.include, "include", "i", "include entry types")
 	persistentFlags.BoolVarP(&c.update.recursive, "recursive", "r", c.update.recursive, "recursive")
+	persistentFlags.BoolVar(&c.update.useBuiltinGit, "use-builtin-git", c.update.useBuiltinGit, "use builtin git")
 
 	return updateCmd
 }
 
 func (c *Config) runUpdateCmd(cmd *cobra.Command, args []string) error {
-	if err := c.run(c.absSourceDir, c.Git.Command, []string{"pull", "--rebase"}); err != nil {
-		return err
+	if c.update.useBuiltinGit {
+		rawSourceDir, err := c.baseSystem.RawPath(c.absSourceDir)
+		if err != nil {
+			return err
+		}
+		repo, err := git.PlainOpen(rawSourceDir)
+		if err != nil {
+			return err
+		}
+		wt, err := repo.Worktree()
+		if err != nil {
+			return err
+		}
+		if err := wt.Pull(&git.PullOptions{
+			RecurseSubmodules: git.DefaultSubmoduleRecursionDepth,
+		}); err != nil && !errors.Is(err, git.NoErrAlreadyUpToDate) {
+			return err
+		}
+	} else {
+		args := []string{
+			"pull",
+			"--rebase",
+			"--recurse-submodules",
+		}
+		if err := c.run(c.absSourceDir, c.Git.Command, args); err != nil {
+			return err
+		}
 	}
 
 	if !c.update.apply {